moment equations update
